+++ components/ving/FormLabel.vue
@@ -1,5 +1,5 @@
 <template>
-    <label v-if="label" :for="id" class="block font-medium text-900 mb-2">{{ label }}</label>
+    <label v-if="label" :for="id" class="block font-medium mb-2">{{ label }}</label>
 </template>
 
 <script setup>