+++ components/ving/UserProfileLink.vue
@@ -1,5 +1,5 @@
 <template>
-    <NuxtLink v-if="user" :to="`/user/${user.props.id}/profile`" v-ripple>
+    <NuxtLink v-if="user" :to="`/user/${user.props.id}/profile`">
         <UserAvatar :user="user" />
     </NuxtLink>
     <Message v-else severity="error">You may need to <i>includeRelated:['user']</i> to view this.</Message>