+++ layouts/default.vue
@@ -1,55 +1,53 @@
 <template>
     <div>
-            <Menubar id="topnav" :model="topNav" class="bg-gray-900 border-noround py-0">
-                <template #start>
-                    <img :src="config.public.site.logoUrl" :alt="config.public.site.name" height="40"
-                        class="mr-0 lg:mr-6">
-                </template>
-                <template #item="{ item, props, hasSubmenu, root }">
-                    <a v-if="hasSubmenu" v-ripple :target="item.target" v-bind="props.action"
-                        class="flex px-6 p-3 lg:px-3 lg:py-2 align-items-center bg-gray-900 font-medium border-round cursor-pointer text-gray-400 hover:text-white hover:bg-gray-800">
-                        <Icon :name="item.icon" class="mr-2" />
-                        <span class="ml-2">{{ item.label }}</span>
-                        <Icon v-if="hasSubmenu" name="pepicons-pop:angle-down" class="ml-2"/>
-                    </a>
-                    <NuxtLink v-else :to="item.to" v-ripple
-                        class="flex px-6 p-3 lg:px-3 lg:py-2 bg-gray-900 align-items-center text-gray-400 hover:text-white hover:bg-gray-800 font-medium border-round cursor-pointer">
-                        <Icon :name="item.icon" class="mr-2" />
-                        <span>{{ item.label }}</span>
-                    </NuxtLink>
-                </template>
+        <Menubar id="topnav" :model="topNav" class="py-0">
+            <template #start>
+                <img :src="config.public.site.logoUrl" :alt="`${config.public.site.name} logo`" :title="config.public.site.name" class="h-10 mr-0 lg:mr-3">
+            </template>
+            <template #item="{ item, props, hasSubmenu, root }">
+                <a v-if="hasSubmenu" :target="item.target" v-bind="props.action"
+                    class="flex items-center px-6 p-3 lg:px-3 lg:py-2 rounded cursor-pointer">
+                    <Icon :name="item.icon" class="mr-2" />
+                    <span class="ml-2">{{ item.label }}</span>
+                    <Icon v-if="hasSubmenu" name="pepicons-pop:angle-down" class="ml-2"/>
+                </a>
+                <NuxtLink v-else :to="item.to"
+                    class="flex items-center px-6 p-3 lg:px-3 lg:py-2 rounded cursor-pointer">
+                    <Icon :name="item.icon" class="mr-2" />
+                    <span>{{ item.label }}</span>
+                </NuxtLink>
+            </template>
 
-                <template #end>
-                    <div class="flex align-items-center gap-2">
-                        <InputGroup class="border-secondary">
-                            <InputGroupAddon class="bg-gray-900 border-primary">
-                                <Icon name="ion:search"/>
-                            </InputGroupAddon>
-                            <InputText placeholder="Search (non-functional)" type="text"
-                                class="w-8rem sm:w-auto bg-gray-900 text-white border-primary" />
-                        </InputGroup>
-                        <SplitButton v-if="currentUser.props?.id" :model="userMenu" text>
-                            <NuxtLink to="/user/settings" v-ripple class="flex align-items-center">
-                                <Avatar :image="currentUser.meta?.avatarUrl" alt="user avatar" shape="circle" />
-                                <span class="text-white ml-2">
-                                    {{ currentUser.meta?.displayName }}
-                                </span>
-                            </NuxtLink>
-                            <template #item="{ item }">
-                                <NuxtLink :to="item.to" v-ripple class="flex p-3 align-items-center">
-                                    <Icon :name="item.icon" class="mr-2" />
-                                    {{ item.label }}
-                                </NuxtLink>
-                            </template>
-                        </SplitButton>
-                        <NuxtLink v-else to="/user/login" v-ripple style="min-width: 120px;"
-                            class="flex p-3 align-items-center text-gray-400 hover:text-white hover:bg-gray-800 no-underline">
-                            <Icon name="fa6-solid:door-open" class="mr-2" />
-                            Sign In
+            <template #end>
+                <div class="flex items-center gap-2">
+                    <DarkModeSelector/>               
+                    <InputGroup>
+                        <InputGroupAddon>
+                            <Icon name="ion:search"/>
+                        </InputGroupAddon>
+                        <InputText placeholder="Search (non-functional)" type="text" class="w-8rem sm:w-auto" />
+                    </InputGroup>
+                    <SplitButton v-if="currentUser.props?.id" :model="userMenu" text>
+                        <NuxtLink to="/user/settings" class="flex items-center">
+                            <Avatar :image="currentUser.meta?.avatarUrl" alt="user avatar" shape="circle" />
+                            <span class="ml-2">
+                                {{ currentUser.meta?.displayName }}
+                            </span>
                         </NuxtLink>
-                    </div>
-                </template>
-            </Menubar>
+                        <template #item="{ item }">
+                            <NuxtLink :to="item.to" class="flex p-3 items-center">
+                                <Icon :name="item.icon" class="mr-2" />
+                                {{ item.label }}
+                            </NuxtLink>
+                        </template>
+                    </SplitButton>
+                    <NuxtLink v-else to="/user/login" class="flex p-3 items-center text-nowrap">
+                        <Icon name="fa6-solid:door-open" class="mr-2" />
+                        Sign In
+                    </NuxtLink>
+                </div>
+            </template>
+        </Menubar>
 
         <SystemWideAlert />
         <div class="px-0 py-1 md:px-4">
@@ -91,4 +89,5 @@ const userMenu = computed(() => {
         out.unshift({ label: 'Admin', to: '/admin', icon: 'prime:user-plus' });
     return out;
 })
+
 </script>