+++ pages/user/[id]/profile.vue
@@ -1,10 +1,10 @@
 <template>
     <Title>User profile for {{ user.meta?.displayName }}</Title>
-    <div v-if="user.props?.id" class="surface-card p-4 border-1 surface-border border-round flex-auto">
+    <div v-if="user.props?.id" class="surface-card p-4 border border-surface rounded flex-auto">
         <div class="text-900 font-semibold text-lg mt-3">User profile for {{ user.meta?.displayName }}</div>
 
         Created on {{ formatDate(user.props?.createdAt) }}
-        <div class="flex gap-5 flex-column-reverse md:flex-row">
+        <div class="flex gap-5 flex-col-reverse md:flex-row">
             <div class="flex-auto">
                 <div class="mb-4">
 
@@ -16,8 +16,8 @@
                 </div>
 
             </div>
-            <div class="flex flex-column align-items-center flex-or">
-                <Avatar :image="user.meta?.avatarUrl" alt="user avatar" class="h-10rem w-10rem" shape="circle" />
+            <div class="flex flex-col items-center flex-or">
+                <Avatar :image="user.meta?.avatarUrl" alt="user avatar" size="xlarge" shape="circle" />
             </div>
         </div>