+++ pages/user/create.vue
@@ -1,8 +1,8 @@
 <template>
     <Title>Create an Account for {{ config.public?.site?.name }}</Title>
-    <div class="w-full lg:w-7 mx-auto">
+    <div class="w-full lg:w-7/12 mx-auto">
         <div class="text-center mb-5">
-            <img :src="config.public.site.logoUrl" :alt="config.public.site.name" height="50" class="mb-3">
+            <img :src="config.public.site.logoUrl" :alt="config.public.site.name" class="inline-block h-16 mb-3">
             <h1 class="text-900 text-3xl font-medium mb-3 mt-0">Create an Account</h1>
             <span class="text-600 font-medium line-height-3">Already have an account?</span>
             <NuxtLink to="/user/login" class="font-medium no-underline ml-2 text-blue-500 cursor-pointer">Sign in
@@ -26,27 +26,24 @@
                     autocomplete="new-password" type="password" :mustMatch="{ field: 'Password', value: newUser.password }"
                     class="mb-3" />
 
-                    <Button type="submit" class="w-full"><span class="text-center w-full"><Icon name="ph:user" class="mr-1"/><span class="vertical-align-middle">Create Account</span></span></Button>
+                    <Button type="submit" class="w-full"><span class="text-center w-full"><Icon name="ph:user" class="mr-1"/><span class="align-middle">Create Account</span></span></Button>
             </Form>
 
             <Divider align="center" class="my-4">
-                <span class="text-600 font-normal text-sm">OR</span>
+                <span >OR</span>
             </Divider>
 
-            <div class="flex justify-content-between">
-                <Button
-                    class="mr-2 w-6 font-medium border-1 surface-border surface-100 py-3 px-2 p-component hover:surface-200 active:surface-300 text-900 cursor-pointer transition-colors transition-duration-150 inline-flex align-items-center justify-content-center">
-                    <Icon name="prime:facebook" class="text-indigo-500 mr-2" size="50px" />
+            <div class="flex justify-between">
+                <Button class="mr-2">
+                    <Icon name="prime:facebook" class="mr-2" size="3rem" />
                     <span>Sign in With Facebook</span>
                 </Button>
-                <Button
-                    class="ml-2 w-6 font-medium border-1 surface-border surface-100 py-3 px-2 p-component hover:surface-200 active:surface-300 text-900 cursor-pointer transition-colors transition-duration-150 inline-flex align-items-center justify-content-center">
-                    <Icon name="prime:github" class="text-black-500 mr-2" size="50px" />
+                <Button>
+                    <Icon name="prime:github" class="mr-2" size="3rem" />
                     <span>Sign in With GitHub</span>
                 </Button>
-                <Button
-                    class="ml-2 w-6 font-medium border-1 surface-border surface-100 py-3 px-2 p-component hover:surface-200 active:surface-300 text-900 cursor-pointer transition-colors transition-duration-150 inline-flex align-items-center justify-content-center">
-                    <Icon name="prime:google" class="text-red-500 mr-2" size="50px" />
+                <Button class="ml-2">
+                    <Icon name="prime:google" class="mr-2" size="3rem" />
                     <span>Sign in With Google</span>
                 </Button>
             </div>